Fix neighbor and edge iterators visiting self loops twice
Fixes #31

For an undirected edges traversal, selfloops would be visited twice
since the same edge is of course linked in both in the outgoing as well
as incoming edge lists.

Fix the code in the Edges iterator, the Neighbors iterator is simply
equivalent, and we make it a wrapper of Edges.

[breaking-change]