Add support for computing dominators
This is an implementation of the engineered ["Simple, Fast Dominance
Algorithm"][0] discovered by Cooper et al.

This algorithm is `O(v^2)`, and therefore has slower theoretical running
time than the Lenguaer-Tarjan algorithm (which is `O(e *
log(v)`). However, Cooper et al found it to be faster in practice on
control flow graphs of up to ~30,000 vertices. Additionally, the
algorithm is much easer to understand than LT and, and easier to
implement as well.

[0]: http://www.cs.rice.edu/~keith/EMBED/dom.pdf