dominators: Implement the Lengauer-Tarjan algorithm
This algorithm runs in **O(|E| log |V|)** time. This is a better theoretical
running time than the "Simple, Fast" algorithm by Cooper (see `simple_fast`),
but has higher start-up costs and memory requirements. The "Simple, Fast" paper
reported the tipping point where the Lengauer-Tarjan algorithm outperformed
theirs at CFGs with >= 30,000 vertices. Take this with a grain of salt, however,
as noted in ["Finding Dominators in Practice" by Loukas Georgiadis, Robert
E. Tarjan, and Renato F. Werneck][1]:

> Even on small instances, however, we did not observe the clear superiority
> of the ["Simple, Fast" algorithm] reported by Cooper et al.,
> which we attribute to their inefficient implementation of [the simple
> Lengauer-Tarjan algorithm].

Note that this is an implementation of the simple version of the Lengauer-Tarjan
algorithm. There is another version that has an even better theoretical running
time of **O(|E| α(|E|, |V|))** where **α** is a functional inverse of
Ackermann's function. However, "Finding Dominators in Practice" found it to be
slower than the simpler version for all graphs they tested:

> Both versions of the Lengauer-Tarjan algorithm (LT [the advanced version])
> and (SLT [the simple version]) are more robust [than the "Simple,
> Engineered" algorithm] on application graphs, and the advantage increases
> with graph size or graph complexity. Among these three, LT was the slowest,
> in contrast with the results reported by Lengauer and Tarjan [30]. SLT and
> [yet another dominators algorithm] were the most consistently fast
> algorithms in practice; since the former is less sensitive to pathological
> instances, we think it should be preferred where performance guarantees
> are important.

[0]: http://www.cs.princeton.edu/courses/archive/spr03/cs423/download/dominators.pdf
[1]: http://jgaa.info/accepted/2006/GeorgiadisTarjanWerneck2006.10.1.pdf