Add an immutable Graph::node_weights() iterator
This adds a new helper function to Graph that allows users to
immutably iterate through all node weights.

This is similar to the existing Graph::node_weights_mut() function.

This fills a tiny gap in the API - in most cases, every mutable accessor is
accompanied by an immutable accessor of the same name. This is so callers from
immutable contexts can use the same API as the corresponding logic but in
an mutable context..