Ensure A* algorithm re-visits nodes when needed
Fix for #378

Given an admissible but inconsistent heuristic (that is, one that for
every node underestimates the distance to the goal, but may overestimate
the distance from a neighbour of to the goal plus the cost of reaching
that neighbour), A* may need to re-visit some non-goal nodes as it finds
shorter paths to reach them.

This commit forces A* to do these re-expansions. It also simplifies the
implementation a bit, removing the unnecessary use of a visit map to
keep track of visited nodes (the tracking is already happening due to
having to keep track of the costs).