Add EdgeIndexable public trait
Just as there is a NodeIndexable public trait that lets users handle the
interaction with node indices there was no corresponding trait for
edges. This commit adds a new public trait EdgeIndexable which mirrors
NodeIndexable but works with edge indices instead. It is then
implemented for StableGraph and frozen (the graphs where there were
NodeIndexable impls and EdgeIndexable made sense).