Organize visit trait implementations (#448)
* Move graph trait impls out of visit

Some graphs had their visit trait implementations in the visit modules,
while others had them in the graph module (and normally it was a mix
between the two).

This commit moves all the impls to the respective graph modules,
to make the organization of the code more consistent.

* Add table listing each graph traits in visit doc

The visit:: traits can be a bit confusing at first. This change to the
visit module documentation adds a quick overview over which graph
implements which trait.

It may come useful when deciding the traits to require for a new generic algorithm.