Address more clippy lints (#449)
* Add clippy config with MSRV

* Allow float comparision with 0

We don't need to use an epsilon as the 0 values are always assigned directly
from a constant.

* Use &self method receiver where possible

The trait_template! macro does not support that, so I added explicit
ignores to clippy's warning.

* Custom bellman_fords output struct

The previous type was overly complex and not too intuitive.

* Address type complexity lints

In most cases I think the type is Ok as-is, so I added an ignore

* Address remaining clippy lints

The needless_range_loop in floid_warshall looks more intuitive using
the explicit range, hence I ignored the lint.