Update `indexmap` to v2
`indexmap` v2.0.0 has just been released, and several prominent crates
in the Rust ecosystem such as `serde_json` are already using it. To
minimize the number of projects for which Cargo would have to download
and build two versions of `indexmap` due to their transitive
dependencies, let's update tyhe version `petgraph` depends on.

`indexmap` v2.0.0 does not bring substantial changes for this project,
but it bumps our MSRV to 1.64. This should not be a concern, given that
0.7.0 (#567) will bump it to 1.65 anyway.