Add element filter adapter
Due to the implementation of several traits on
graph-refs, it is not possible to nest a
`NodeFiltered` within an `EdgeFiltered`, and
vice-versa. The inner graph must be a GraphRef,
which imposes lifetime constraints that can't be
resolved in many cases (e.g. as a field within
a struct, or as a return value from a function).

This PR adds the `ElementFiltered` adapter
address this problem, and a `NestedFilter` structure
for the most common way to combine a node filter and
an edge filter.