feat: Add LowerHex and UpperHex impls for Dot (#687)
Hello!

Thank you for this library, it's very nice!

I'm using `petgraph` to represent some data with numeric edge and node
weights. I'm sometimes using the `Dot` struct to display these graphs,
and would like to be able to use `{:x}` and `{:X}` formatting for node
and edge labels.

This PR just implements `LowerHex` and `UpperHex` on `Dot` in the style
of the existing `Display` and `Debug` implementations.