ci: Require conventional commits tag in PR titles (#734)
This is part of the CI improvements listed in #725.

Adds a check when a PR is opened or modified to ensure it's title
contains a [conventional
commits](https://www.conventionalcommits.org/en/v1.0.0/) tag, e.g.
```yml
# Some new features and bug fixes
feat: Add this algorithm
fix(component): Don't panic
# A PR that introduces a semver breaking change
fix!: Refactor all the traits
```
When a PR introduces breaking changes (that will require a breaking
semver bump) we also require a `BREAKING CHANGE: ...` line in the PR
body describing what the changes required in user code.

When the check fails, the bot posts a help message indicating what needs
to be fixed.

The list of valid tags are: `feat, fix, docs, style, refactor, perf,
test, ci, chore, revert`.
`rfc` is an invalid tag with a custom error message, useful when opening
draft PRs.
I added a pull request template with these instructions.

The main goal of this requirement is to automate changelog generation
with [release-plz](https://release-plz.dev/). The tags get grouped into
categories in the changelog, and breaking changes are automatically
picked when deciding the version bump.

This also integrates with the semver checks added in #692. The check
fails on PRs with breaking changes unless they are marked as breaking in
title.

---

The code here is based on the reusable workflow I implemented in
[CQCL/hugrverse-actions](https://github.com/CQCL/hugrverse-actions?tab=readme-ov-file#pr-title).