feat: Add Shortest Path Faster Algorithm Implementation (#686)
SPFA (Shortest Path Faster Algorithm) is the improved variation of
Bellman-Ford.

### Time Complexity: 
- Average Time Complexity: `O(|E|) `, but this bound has not been
approved yet.
- Worstcase Time Complexity: `O(|V|*|E|)`

### Space complexity
- Space complexity: `O(V)`

###
[Wikipedia](https://en.wikipedia.org/wiki/Bellman%E2%80%93Ford_algorithm#CITEREFMoore1959)
says:
A variation of the Bellman–Ford algorithm described by [Moore
(1959)](https://en.wikipedia.org/wiki/Bellman%E2%80%93Ford_algorithm#CITEREFMoore1959),
reduces the number of relaxation steps that need to be performed within
each iteration of the algorithm. If a vertex v has a distance value that
has not changed since the last time the edges out of v were relaxed,
then there is no need to relax the edges out of v a second time. In this
way, as the number of vertices with correct distance values grows, the
number whose outgoing edges that need to be relaxed in each iteration
shrinks, leading to a constant-factor savings in time for [dense
graphs](https://en.wikipedia.org/wiki/Dense_graph). This variation can
be implemented by keeping a collection of vertices whose outgoing edges
need to be relaxed, removing a vertex from this collection when its
edges are relaxed, and adding to the collection any vertex whose
distance value is changed by a relaxation step. In China, this algorithm
was popularized by Fanding Duan, who rediscovered it in 1994, as the
"shortest path faster
algorithm".[[6]](https://en.wikipedia.org/wiki/Bellman%E2%80%93Ford_algorithm#cite_note-6)

### About implementation
I used a stack instead of a queue to improve performance. In this
regard, there is a slight discrepancy with the classic description of
the "SPFA" algorithm, but this option was also tested on random tests
against Bellman-Ford
[here](https://github.com/starovoid/graphalgs/blob/main/src/shortest_path/spfa.rs).