ci: Use `release-plz` for automated changelog generation and publication (#735)
Setups [`release-plz`](https://release-plz.dev/) to run on each push to
`master`.
This has two different flows:

### Release PR (disabled until after the `0.8.0` release)

When a PR is merged, it analyses all the commits since the last release
and creates a draft PR with the automatically generated changelog.
Commits are grouped into categories using conventional commit tags (see
#734). I ported the existing `RELEASES.rst` to markdown, so updates can
be done from there.

It also runs
[`cargo-semver-checks`](https://crates.io/crates/cargo-semver-checks) to
decide if the next release should be a semver-breaking bump, and updates
the package version.
Any merged PR marked as breaking (#734) will also cause a breaking
version bump.

### Package publication

When a change is merged that bumps the package version `release-plz`
will create a new github release with the appropriate changelog and a
list of contributors, and push the new version to crates.io.
(I haven't yet added the crates token, will do once we verify this
action is correctly configured, before the new release gets merged).


drive-by: cherry-pick new clippy lint fixes from #753

---------

Co-authored-by: starovoid <prototyperailgun@gmail.com>