ci: Enable automatic CHANGELOG generation (#761)
This will enable automatic release PR generation by `release-plz`.
This includes a semver-checks run to ensure that no breaking checks get
released in a minor version bump, and CHANGELOG generation based on the
commits merged since the last release, grouped by their conventional
commits tag (#734).

PRs marked as breaking (by adding a `!` after the tag) will be marked as
so in the CHANGELOG, and they will force a major version bump.

It is possible to modify the created PR before merging as needed; either
to reword parts of the changelog or to override the version bump.