feat(parser): allow parsing graphs from Dot/Graphviz files (#653)
This is a PR that adds support for Dot/Graphviz parsing. Parsing can
happen both at runtime and at compile time (i.e. there are functions and
macros).

It fixes #446 and #466 .

Internally, it uses the `dot_parser` crate
(https://crates.io/crates/dot-parser), the implementation is simply a
thin wrapper over that crate. This wrapper is a trait (`ParseFromDot`),
introduced in a new module `crate::dot::dot_parser`. I also added a
feature `dot_parser` to gate this module.

The modifications do *not* introduce breaking changes.

TODO List (this is just what I have on the top of my head right now,
feel free to edit if you think of other items):
- [x] Add more tests.
- [x] Implement `ParseFromDot` for existing graph types.
- [x] Discuss on exposed (and re-exported) items.
- [x] Improve the documentation. (If need be, I'll also improve the doc
on the `dot_parser` crate side).