feat: Implement `DataMap` for `GraphMap` graphs (#776)
The `DataMap` implementation is the only trait missing to make
`GraphMap` graphs usable with `isomorphism` functions.

The `Vec<usize>` return type of isomorphism matchings is slightly
confusing for `GraphMap`, but I think it's fine to include this PR by
itself. In the future a specific "MatchResolver" or similar could always
be added for `GraphMap`s that turns the `Vec<usize>` (which is using the
order of `GraphMap::nodes`) into a `G0::N ==> G1::N` mapping.