fix: Ford Fulkerson sometimes Panics on StableGraphs (#793)
This Resolves #792 and should fix the problem as a whole for Ford
Fulkerson.

As explained in #792 , the previous implementation assumed continuous
edge indices (that is, without gaps), which is not necessarily the case
with StableGraphs. This implementation will now use a bit more memory by
creating and returning a vector of `edge_bound()` size, instead of
`edge_count()`.

This should not change the behaviour of the algorithm for any other
graph class except for StableGraphs.

This new behavior is also reflected in the new docstrings of the
algorithm. A test was also added to check that the issue does not
reappear.

In the future, one could think about returning a `HashMap` with
`edge_count()` many entries instead of a `Vec`, but this would change
the API.