feat: Add algorithm to find bridge edges (#590)
Derived from #473 (written by @starovoid).

Here's a list of changes that we (@gmorenz and myself) made:
- Used iteration instead of recursion; with the original version we hit
a stack overflow.
- Tried to make variable names self documenting (e.g. `fup` to
`earliest-backedge`)
- Change the output type from `Vec<G::NodeId, G::NodeId)>` to `impl
Iterator<Item = G::EdgeRef>`

The original code also had an algorithm to find articulation points
(node variant of bridge edges). To keep the pull request small we only
added the bridge algorithm. We would be happy to include the other
algorithm in another pull request. As a result we named the module
`bridges` instead of `bridges_and_ap`, we think this is probably a
better (more succinct) name anyways.

---------

Co-authored-by: xylqn7 <xylqn7@morel.SafeAI-Guest.local>
Co-authored-by: xylqn7 <xylqn7@morel.lan>