docs: Add examples to minimum spanning tree functions (#808)
This PR simply adds examples (really the same one for both) to the
minimum spanning tree functions (the default one and the prim variant).
Adding these makes them more consistent with the rest of the doc strings
for other functions, since most of them include examples of how to use
them.

This seems particularly useful, since in case of computing msts with
petgraph, one needs to use `from_elements`, which is mentioned in the
docstring already, but it might not be clear how to actually use
`from_elements` to obtain a graph.

---------

Co-authored-by: Egor Starovoitov <52821033+starovoid@users.noreply.github.com>