docs: Update docs.rs (#807)
This PR updates the docs.rs by mainly replacing the existing example by
several other ones, which go into more detail. It also restructures the
main docs while leaving most of the information the same by adding an
overview section.

The goal of this is to offer a more smooth introduction into petgraph,
by increasing the clarity of the docs and adding a usage section to help
people that are new to rust get started. With a similar intent, more
examples are added to show off some of petgraphs' features.

Among other things, this resolves #804. 

For a preview of the new docs, just run
```
cargo doc --open
```
after cloning this PR branch.

Note that for the revamp I took inspiration from the documentation style
of several of [@BurntSushi 's crates](https://github.com/burntsushi).

---------

Co-authored-by: Egor Starovoitov <52821033+starovoid@users.noreply.github.com>