docs: Fix link and mention Dfs and Bfs as special case in examples (#816)
I just noticed that the link for the example `Using algorithms on
graphs` in the docs does not work properly because there was a typo.

While at it, I also added a sentence about `Dfs` and `Bfs` not being
found in the algo module and instead in visit.
Similarly, I added links to the docstring of the `Topo` struct which was
referring to other functions/algs without linking to them.