docs: Add link to multigraph definition to isomorphism algos (#824)
This resolves #600.

There seems to have been some confusion on what the term Multigraph
refers to. This should clear up the issue. The further discussion in the
issue seems to be somewhat chaotic due to the missing definition of
multigraph and possibly mixing it up with disconnected graphs.

To address this issue, I added links to definitions of multigraphs to
the respective docstrings.

Once new PR / Issue templates are in place (#811), I will create an
issue that isomorphism functionality should support multigraphs.