docs: Fix auxiliary space (and time) complexity of bron-kerbosch (#825)
I think the auxiliary space and time complexity as recently updated in
#815, are not quite right for Bron-Kerbosch.

For one, the auxiliary space complexity cannot be right. Since, we
return a `Vec` with `HashSet`, the auxiliary space has to be at least
equal to the number of maximal cliques in the graph. [Moon and Moser
showed](https://link.springer.com/article/10.1007/BF02760024) that this
can be up to `3 ^ (|V| / 3)` many. Thus, I suggest introducing `k` as a
variable to denote the number of maximal cliques in the graph.

The time complexity might also be able to be written in terms of `k`,
however this would probably also depend on this implementation of
Bron-Kerbosch with pivoting. Instead one could write something like:

```
/// * Time complexity: **O(p(|V|)k)**
```
For some polynomial `p`, which makes clear that the runtime is not
always exponential in the number of vertices, but is merely dominated by
the number of maximal cliques in the graph.