ci: Speedup miri by using nextest (#810)
[Nextest](https://github.com/nextest-rs/nextest) is an alternative test
runner for rust projects which should speedup the tests, which is
particularly interesting for the miri test run, since that one takes
very long in comparison to the others. Additionally, nextest slightly
changes the output when running tests and better highlights slow tests
for example.

I also included the changes from PR #809 already. If those should not be
desireable, I can adjust the PR accordingly. However, the speedup should
be even more noticeable if the test as suggest in PR #809 is removed,
since most notably, nextest runs some tests in parallel, which is not
very helpful if there are single tests which take extraordinarily long.

There is an [integration with
miri](https://nexte.st/docs/integrations/miri/) which is used for this
PR, for which there is one thing to note however:

"Note, however, that `cargo miri test` is able to detect data races
where two tests race on a shared resource. Miri with nextest will not
detect such races."

I thought that covering such cases is not important for `petgraph`, but
I am open to discussion on that.

---------

Co-authored-by: Agustín Borgna <agustinborgna@gmail.com>