docs: Revamp CONTRIBUTING.md (#833)
This PR revamps the CONTRIBUTING.md. First of all, it switches from .rst
to .md, since the latter seems more modern :star2:

Second, it updates the information in the contribution guide since an
up-to-date contribution guide should make it more attractive to
contribute to the project :) I tried to keep the information there was
before and add some new things or update the things that were outdated
^^

In particular, regarding the different kinds of PRs and their
requirements, I left the old section in the contribution guide as-is,
since I plan to include that information in the PR templates as
mentioned in https://github.com/petgraph/petgraph/issues/811.

For one, this resolves https://github.com/petgraph/petgraph/issues/827,
but also checks the last box in
https://github.com/petgraph/petgraph/issues/725.

---------

Co-authored-by: Egor Starovoitov <52821033+starovoid@users.noreply.github.com>