refactor: replace uses of 'core::iter::zip' with 'crate::util::zip'
The function 'zip' was introduced in #747 as a part of adding `no_std`
support. However, there is already such a function in `core::iter`.
The MSRV stated in the description of #747 was 1.64 at minimum, but the
function in `core` is stable since Rust 1.59. Thus, it is an unnecessary
duplication and can be removed without consequences.