feat: make Csr::from_sorted_edges generic over edge type and properly increase edge_count in Csr::from_sorted_edges (#861)
This PR continues the work done in #783. Thus, this resolves #588.

The trait bound on the `Csr` to be `Directed` for `from_sorted_edges`
was lifted. Thus, it was noticed that at no point in
`from_sorted_edges`, the `edge_count` is increased. This is only
relevant for `Undirected` graphs, as `Directed` graphs use their
`self.column.len()` as the `edge_count()` under the hood.

---------

Co-authored-by: Jan Baudisch <jan.baudisch@uni-ulm.de>