feat: Add `map_owned` and `filter_map_owned` for `Graph` and `StableGraph` (#863)
This PR continues the work done in PR #794. Thus, it resolves #542.

It adds `map_owned` and `filter_map_owned` methods for both `Graph` and
`StableGraph` which take ownership of the respective graphs and their
respective associated data in comparison to the `map` and `filter_map`
counterparts which only take `&` references.

Appropriate tests for this have been added. If desired, I can also add
quickchecks, but this did not seem necessary.

---------

Co-authored-by: Pete Hayes <pete@hayes.id.au>