fix: Algos don't work on `UndirectedAdaptor` (#870) (#871)
Closes #870

The `IntoEdges` implementation for `UnidirectedAdaptor` does not produce
the same set of edges references as an equivalently-constructed
undirected graph. This prevents search and traversal algorithms from
working if the traversal or search would require traversing an adapted
edge whose underlying directed edge is "incoming" (which would prevent
traversal in the directed case).

The underlying reason is that we currently, correctly concatenate
incoming and outgoing edges from the underlying directed graph but
they're concatenated with the same orientation which produces a set of
redundant `EdgeReferences`-- making the adapted graph equivalent to the
underlying directed graph as far as the `IntoEdges` implementation is
concerned.

The `IntoEdges` implementation for undirected graphs produces two
`EdgeReference`s for each graph edge-- one for each orientation. This is
what allows traversal of edges in either direction.

The solution is to make the `UndirectedAdaptor`'s `IntoEdges` behavior
match that of an undirected graph by reversing the edges with `Incoming`
orientation. That way there will be two `EdgeReferences` for each edge
as there are in undirected graphs.

<!--
  -- Thanks for contributing to `petgraph`! 
  --
-- We require PR titles to follow the Conventional Commits
specification,
-- https://www.conventionalcommits.org/en/v1.0.0/. This helps us
generate
  -- changelogs and follow semantic versioning.
  --
  -- Start the PR title with one of the following:
  --  * `feat:` for new features
  --  * `fix:` for bug fixes
  --  * `refactor:` for code refactors
  --  * `docs:` for documentation changes
  --  * `test:` for test changes
  --  * `perf:` for performance improvements
  --  * `revert:` for reverting changes
  --  * `ci:` for CI/CD changes
-- * `chore:` for changes that don't fit in any of the above categories
  -- The last two categories will not be included in the changelog.
  --
-- If your PR includes a breaking change, please add a `!` after the
type
-- and include a `BREAKING CHANGE:` line in the body of the PR
describing
  -- the necessary changes for users to update their code.
  --
  -->

---------

Co-authored-by: Paul Jakob Schroeder <paul.schroeder@tangramvision.com>