docs: Specify iteration order for neighbors and edges and their variants (#790)
This should fix https://github.com/petgraph/petgraph/issues/637.

I tried my best to deduce the iteration order from the code and various
tests. It should also be considered, whether one actually wants to "fix"
the iteration order in the docs of the respective methods. That is,
whether there should be a guarantee for the user that this iteration may
not change in the future.

Another possibility would be to add the iteration order, as suggested in
this PR with a warning that this iteration might change in the future.

On a similar note, would it make sense to create tests that check
whether this guarantee holds, to make sure that, also in the future, the
docs of the methods actually reflect the reality?